Add a spawner strategy - Equip balancing - Bugfixes with storage and UI clicks - Remove enderchest

+++ src/main/java/me/neoblade298/neorogue/equipment/abilities/Brand.java
@@ -1,6 +1,4 @@
 package me.neoblade298.neorogue.equipment.abilities;
-import me.neoblade298.neorogue.equipment.SessionEquipment;
-
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Particle;
@@ -19,6 +17,7 @@ import me.neoblade298.neorogue.equipment.EquipmentInstance;
 import me.neoblade298.neorogue.equipment.EquipmentProperties;
 import me.neoblade298.neorogue.equipment.EquipmentProperties.PropertyType;
 import me.neoblade298.neorogue.equipment.Rarity;
+import me.neoblade298.neorogue.equipment.SessionEquipment;
 import me.neoblade298.neorogue.player.inventory.GlossaryTag;
 import me.neoblade298.neorogue.session.fight.DamageStatTracker;
 import me.neoblade298.neorogue.session.fight.DamageType;
@@ -40,7 +39,7 @@ public class Brand extends Equipment {
 		aoe = TargetProperties.radius(4, false, TargetType.ENEMY);
 	private static final Circle circ = new Circle(aoe.range);
 	private int burn;
-	private double damage;
+	private int damage;
 	private static final ParticleContainer pc = new ParticleContainer(Particle.CRIMSON_SPORE).count(50).spread(0.3, 0.3).offsetY(2),
 		fill = new ParticleContainer(Particle.LAVA),
 		edges = new ParticleContainer(Particle.FLAME);
